product fixes

+++ app/graphql/types/products/update_product_attributes_type.rb
@@ -3,7 +3,7 @@ module Types
         class UpdateProductAttributesType < Types::BaseInputObject
             argument :id, ID, required: true
             argument :name, String, required: false
-            argument :price, Int, required: false
+            argument :price, Float, required: false
             argument :description, String, required: false
             argument :category_id, ID, required: false
             argument :pizza_border_id, ID, required: false
